+++ store/store.go
@@ -1,16 +1,17 @@
 package store
 
 import (
+	dbm "github.com/cosmos/cosmos-db"
 	"github.com/tendermint/tendermint/libs/log"
-	dbm "github.com/tendermint/tm-db"
 
 	"github.com/cosmos/cosmos-sdk/store/cache"
+	"github.com/cosmos/cosmos-sdk/store/metrics"
 	"github.com/cosmos/cosmos-sdk/store/rootmulti"
 	"github.com/cosmos/cosmos-sdk/store/types"
 )
 
-func NewCommitMultiStore(db dbm.DB, logger log.Logger) types.CommitMultiStore {
-	return rootmulti.NewStore(db, logger)
+func NewCommitMultiStore(db dbm.DB, logger log.Logger, metricGatherer metrics.StoreMetrics) types.CommitMultiStore {
+	return rootmulti.NewStore(db, logger, metricGatherer)
 }
 
 func NewCommitKVStoreCacheManager() types.MultiStorePersistentCache {