+++ x/auth/testutil/app_config.go
@@ -1,15 +1,15 @@
 package testutil
 
 import (
-	_ "github.com/cosmos/cosmos-sdk/x/auth"            // import auth as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config"  // import auth as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/auth/vesting"    // import vesting as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/bank"            // import bank as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/consensus"       // import consensus as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/feegrant/module" // import feegrant as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/genutil"         // import genutil as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/params"          // import params as a blank for app wiring
-	_ "github.com/cosmos/cosmos-sdk/x/staking"         // import staking as a blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/auth"            // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/auth/tx/config"  // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/auth/vesting"    // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/bank"            // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/consensus"       // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/feegrant/module" // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/genutil"         // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/params"          // import as blank for app wiring
+	_ "github.com/cosmos/cosmos-sdk/x/staking"         // import as blank for app wiring
 
 	"cosmossdk.io/core/appconfig"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
@@ -43,23 +43,13 @@ var AppConfig = appconfig.Compose(&appv1alpha1.Config{
 				AppName: "AuthApp",
 				BeginBlockers: []string{
 					stakingtypes.ModuleName,
-					authtypes.ModuleName,
-					banktypes.ModuleName,
 					genutiltypes.ModuleName,
 					feegrant.ModuleName,
-					paramstypes.ModuleName,
-					consensustypes.ModuleName,
-					vestingtypes.ModuleName,
 				},
 				EndBlockers: []string{
 					stakingtypes.ModuleName,
-					authtypes.ModuleName,
-					banktypes.ModuleName,
 					genutiltypes.ModuleName,
 					feegrant.ModuleName,
-					paramstypes.ModuleName,
-					consensustypes.ModuleName,
-					vestingtypes.ModuleName,
 				},
 				InitGenesis: []string{
 					authtypes.ModuleName,